Comentários dos patterns strategy, observer e mediator.
