+++ src/com/examples/mediator/Motor.java
@@ -1,5 +1,9 @@
 package com.examples.mediator;
 
+/**
+ * A classe motor contém apenas seu status (se está ligado ou não). Os métodos apenas atribuem os status
+ * e devolvem o status para uso pelas outras classes.
+ */
 public class Motor {
 
     private boolean motorLigado = false;
