+++ src/com/examples/mediator/ParteEletrica.java
@@ -1,12 +1,18 @@
 package com.examples.mediator;
 
+/**
+ * A parte elétrica do carro devolve essencialmente se a mesma está habilitada ou não. Por "habilitada"
+ * definimos que a carga restante deve estar acima de 5.0. Para fins do exemplo, colocamos no construtor
+ * um valor pré-deifnido.
+ */
+
 public class ParteEletrica {
 
     public double cargaRestante;
 
     public ParteEletrica()
     {
-        cargaRestante = 4.0;
+        cargaRestante = 8.0;
     }
 
     public boolean parteEletricaHabilitada()
