--- src/com/examples/observer/IAlarme.java
@@ -1,6 +1,0 @@
-package com.examples.observer;
-
-public interface IAlarme {
-
-    public void alarme();
-}
