+++ src/com/examples/observer/interfaces/IAlarme.java
@@ -1,0 +1,11 @@
+package com.examples.observer.interfaces;
+
+/**
+ * A interface IAlarme possui um único método, que  é o método a ser utilizado caso haja uma mensagem
+ * do observador (no caso SistemaAlarme)
+ */
+
+public interface IAlarme {
+
+    void alarme();
+}
