--- src/com/examples/strategy/IAtaque.java
@@ -1,7 +1,0 @@
-package com.examples.strategy;
-
-
-public interface IAtaque {
-    String atacar();
-    void defineInimigo(String inimigo);
-}
